users_interests/pyproject.toml
[project]
name = "users-interests"
version = "0.1.0"
description = ""
authors = [
    {name = "Juvern",email = "1551120+juvern@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
